+++ downward/reports/__init__.py
@@ -25,8 +25,9 @@ from fnmatch import fnmatch
 import logging
 
 from lab import reports
-from lab import tools
+from lab.reports import markup
 from lab.reports import Attribute, Report, geometric_mean
+from lab import tools
 
 
 class PlanningReport(Report):
@@ -176,6 +177,11 @@ class PlanningReport(Report):
             run.get("node", "<attribute 'node' missing>")
             for run in self.runs.values()}
 
+    def _format_unexplained_errors(self, errors):
+        return markup.escape(errors).replace(
+            '\\n', '\\\\').replace(
+            ' ', markup.ESCAPE_WHITESPACE)
+
     def _get_warnings_text_and_table(self):
         """
         Return a :py:class:`Table <lab.reports.Table>` containing one line for
@@ -197,8 +203,14 @@ class PlanningReport(Report):
             if error_message:
                 logging.error(error_message)
                 num_unexplained_errors += 1
+                run_dir = run['run_dir']
                 for attr in self.ERROR_ATTRIBUTES:
-                    table.add_cell(run['run_dir'], attr, run.get(attr, '?'))
+                    value = run.get(attr, '?')
+                    if attr == 'unexplained_errors':
+                        value = self._format_unexplained_errors(value)
+                        # Use formatted value as-is.
+                        table.cell_formatters[run_dir][attr] = reports.CellFormatter()
+                    table.add_cell(run_dir, attr, value)
 
         if num_unexplained_errors:
             logging.error(