+++ docs/news.rst
@@ -1,7 +1,7 @@
 News
 ====
 
-v5.2 (unreleased)
+v5.2 (2020-01-07)
 -----------------
 
 Lab
@@ -12,7 +12,7 @@ Downward Lab
 ^^^^^^^^^^^^
 * Preserve line breaks for error logs in tables.
 * If an error log in a table has more than 100 lines, omit surplus lines from the middle of the log.
-* Always the print number of runs with unexplained errors when generating any type of report.
+* Always print the number of runs with unexplained errors when generating any type of report.
 
 
 v5.1 (2019-12-10)